// src/grammar/gen/src/grammar/GeometryGrammarListener.ts
export interface GeometryGrammarListener extends ParseTreeListener {
  /**
   * Enter a parse tree produced by `GeometryGrammarParser.command`.
   * @param ctx the parse tree
   */
  enterCommand?: (ctx: CommandContext) => void;
  /**
   * Exit a parse tree produced by `GeometryGrammarParser.command`.
   * @param ctx the parse tree
   */
  exitCommand?: (ctx: CommandContext) => void;

  /**
   * Enter a parse tree produced by `GeometryGrammarParser.drawCmd`.
   * @param ctx the parse tree
   */
  enterDrawCmd?: (ctx: DrawCmdContext) => void;
  /**
   * Exit a parse tree produced by `GeometryGrammarParser.drawCmd`.
   * @param ctx the parse tree
   */
  exitDrawCmd?: (ctx: DrawCmdContext) => void;

  /**
   * Enter a parse tree produced by `GeometryGrammarParser.shape`.
   * @param ctx the parse tree
   */
  enterShape?: (ctx: ShapeContext) => void;
  /**
   * Exit a parse tree produced by `GeometryGrammarParser.shape`.
   * @param ctx the parse tree
   */
  exitShape?: (ctx: ShapeContext) => void;

  /**
   * Enter a parse tree produced by `GeometryGrammarParser.sphere`.
   * @param ctx the parse tree
   */
  enterSphere?: (ctx: SphereContext) => void;
  /**
   * Exit a parse tree produced by `GeometryGrammarParser.sphere`.
   * @param ctx the parse tree
   */
  exitSphere?: (ctx: SphereContext) => void;

  /**
   * Enter a parse tree produced by `GeometryGrammarParser.box`.
   * @param ctx the parse tree
   */
  enterBox?: (ctx: BoxContext) => void;
  /**
   * Exit a parse tree produced by `GeometryGrammarParser.box`.
   * @param ctx the parse tree
   */
  exitBox?: (ctx: BoxContext) => void;

  /**
   * Enter a parse tree produced by `GeometryGrammarParser.pyramid`.
   * @param ctx the parse tree
   */
  enterPyramid?: (ctx: PyramidContext) => void;
  /**
   * Exit a parse tree produced by `GeometryGrammarParser.pyramid`.
   * @param ctx the parse tree
   */
  exitPyramid?: (ctx: PyramidContext) => void;

  /**
   * Enter a parse tree produced by `GeometryGrammarParser.size`.
   * @param ctx the parse tree
   */
  enterSize?: (ctx: SizeContext) => void;
  /**
   * Exit a parse tree produced by `GeometryGrammarParser.size`.
   * @param ctx the parse tree
   */
  exitSize?: (ctx: SizeContext) => void;

  /**
   * Enter a parse tree produced by `GeometryGrammarParser.radius`.
   * @param ctx the parse tree
   */
  enterRadius?: (ctx: RadiusContext) => void;
  /**
   * Exit a parse tree produced by `GeometryGrammarParser.radius`.
   * @param ctx the parse tree
   */
  exitRadius?: (ctx: RadiusContext) => void;

  /**
   * Enter a parse tree produced by `GeometryGrammarParser.color`.
   * @param ctx the parse tree
   */
  enterColor?: (ctx: ColorContext) => void;
  /**
   * Exit a parse tree produced by `GeometryGrammarParser.color`.
   * @param ctx the parse tree
   */
  exitColor?: (ctx: ColorContext) => void;

  /**
   * Enter a parse tree produced by `GeometryGrammarParser.position`.
   * @param ctx the parse tree
   */
  enterPosition?: (ctx: PositionContext) => void;
  /**
   * Exit a parse tree produced by `GeometryGrammarParser.position`.
   * @param ctx the parse tree
   */
  exitPosition?: (ctx: PositionContext) => void;

  /**
   * Enter a parse tree produced by `GeometryGrammarParser.rotation`.
   * @param ctx the parse tree
   */
  enterRotation?: (ctx: RotationContext) => void;
  /**
   * Exit a parse tree produced by `GeometryGrammarParser.rotation`.
   * @param ctx the parse tree
   */
  exitRotation?: (ctx: RotationContext) => void;
}